+++ covalent_ec2_plugin/assets/infra/variables.tf
@@ -2,29 +2,25 @@
 #
 # This file is part of Covalent.
 #
-# Licensed under the GNU Affero General Public License 3.0 (the "License").
-# A copy of the License may be obtained with this software package or at
+# Licensed under the Apache License 2.0 (the "License"). A copy of the
+# License may be obtained with this software package or at
 #
-#      https://www.gnu.org/licenses/agpl-3.0.en.html
+#     https://www.apache.org/licenses/LICENSE-2.0
 #
-# Use of this file is prohibited except in compliance with the License. Any
-# modifications or derivative works of this file must retain this copyright
-# notice, and modified files must contain a notice indicating that they have
-# been altered from the originals.
-#
-# Covalent is distributed in the hope that it will be useful, but WITHOUT
-# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
-# FITNESS FOR A PARTICULAR PURPOSE. See the License for more details.
-#
-# Relief from the License may be granted by purchasing a commercial license.
+# Use of this file is prohibited except in compliance with the License.
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
-variable "name" {
-  default     = "covalent-executor-ec2"
+variable "prefix" {
+  default     = ""
   description = "Name used to prefix AWS resources"
 }
 
-variable "aws_region" {
-  default     = "us-east-1"
+variable "region" {
+  default     = ""
   description = "Region where resources for the EC2 plugin are deployed"
 }
 
@@ -84,6 +80,6 @@ variable "key_file" {
 }
 
 variable "covalent_version" {
-  default     = "0.202.0"
+  default     = ""
   description = "Covalent version to install on the EC2 instance"
 }